fix: Improve clear button input focus management + fix collection state bug
