+++ lib/algolia/useHits.ts
@@ -3,6 +3,7 @@ import { useSearchStateContext } from './context';
 
 interface UseHits<Hit = any> {
    isLoaded: boolean;
+   isSearching: boolean;
    numberOfHits: number;
    hits: Hit[];
 }
@@ -12,9 +13,16 @@ export function useHits<Hit = any>(): UseHits<Hit> {
    const result = React.useMemo<UseHits>(() => {
       return {
          isLoaded: state.isLoaded,
+         isSearching: state.isSearching,
          numberOfHits: state.numberOfHits || state.hits.allIds.length,
          hits: state.hits.allIds.map((id) => state.hits.byId[id]),
       };
-   }, [state.hits.allIds, state.hits.byId, state.isLoaded, state.numberOfHits]);
+   }, [
+      state.hits.allIds,
+      state.hits.byId,
+      state.isLoaded,
+      state.isSearching,
+      state.numberOfHits,
+   ]);
    return result;
 }
